+++ app/src/main/java/com/example/chat/application/chatexample/ChatActivity.java
@@ -161,7 +161,7 @@ public class ChatActivity extends AppCompatActivity
 
     //main and sandbox
 
-    private static String TOKEN = "58ee344c55094c3e827a0c1f04619a71";
+    private static String TOKEN = "869221a3923f49879ecd38824f7d787e";
     private static String ssoHost = BaseApplication.getInstance().getString(R.string.ssoHost);
     private static String serverName = "chat-server";
 
@@ -666,7 +666,7 @@ public class ChatActivity extends AppCompatActivity
                 new RequestThreadInfo.Builder(TEST_THREAD_ID)
                                                 .name("Chat sample thread") // required. if not set, thread name will set to null
 //                                                .metadata("{}") // required. if not set, thread metadata will set to null
-                                                .image("http://sandbox.pod.ir:8443/nzh/drive/downloadImage?hash=WNRI5BHLR4SJCQIE") // required. if not set, thread image will set to null
+                                                .image("https://podspace.pod.ir/nzh/drive/downloadImage?hash=ELJIHZN9NP37ZIDA") // required. if not set, thread image will set to null
                         .description("this is test description updated on " + new Date().toString()) // required. if not set, thread name will set to null
 //                                                .setUploadThreadImageRequest(requestUploadImage) // set when you wanna upload thread image
 //                                                .setUserGroupHash(TEST_THREAD_HASH) // set when you wanna upload thread image
@@ -753,6 +753,7 @@ public class ChatActivity extends AppCompatActivity
     private void downloadWithGlide() {
 
         try {
+            String url = "https://podspace.pod.ir/nzh/drive/downloadImage?hash=V61H4MEOY488X7KJ";
 
             RequestOptions requestOptions = new RequestOptions();
 
@@ -778,9 +779,11 @@ public class ChatActivity extends AppCompatActivity
 
     }
 
-    String url= "https://core.pod.ir/nzh/image?imageId=222808&hashCode=16c3cd2b93f-0.527719303638482";;
-    String imageHashCode = "16feadc7bf1-0.5248624596970302";
+
     private void downloadFile() {
+
+
+        String url = "https://core.pod.ir/nzh/image?imageId=222808&hashCode=16c3cd2b93f-0.527719303638482";
 
         //image
 //        long imageId = 404143;
@@ -790,7 +793,7 @@ public class ChatActivity extends AppCompatActivity
 
 
 //        String imageHashCode = "17024463974-0.5588549950419339";
-
+        String imageHashCode = "16feadc7bf1-0.5248624596970302";
 
         String fileHashCode = "17077360d4b-0.2366487166443898";
 
@@ -801,11 +804,11 @@ public class ChatActivity extends AppCompatActivity
         RequestGetFile requestGetFile = new RequestGetFile.Builder(fileId, fileHashCode, true).build();
 
 
-        RequestGetPodSpaceFile rePod = new RequestGetPodSpaceFile.Builder("SN9HYXS4OO9UL4TB")
+        RequestGetPodSpaceFile rePod = new RequestGetPodSpaceFile.Builder("GAUPE1ZRK76VOXBS")
                 .build();
 
         RequestGetPodSpaceImage rePodImage = new RequestGetPodSpaceImage
-                .Builder(imageHashCode)
+                .Builder("613Q7WCCEXZ1DGY5")
 //                .setCrop(true)
                 .setQuality(0.45f)
                 .build();
@@ -1492,11 +1495,12 @@ public class ChatActivity extends AppCompatActivity
                 getUri(),
                 TextMessageType.Constants.POD_SPACE_PICTURE) // constructor
                 .description("test file message")
+//                                .systemMetadata(getMetaData())
                 .setUserGroupHash(TEST_THREAD_HASH)
-                                .setImageHc("100")
-                                .setImageWc("100")
-                                .setImageXc("1")
-                                .setImageYc("1")
+//                                .setImageHc("100")
+//                                .setImageWc("100")
+//                                .setImageXc("1")
+//                                .setImageYc("1")
                 .build();
 
         fileUnique[0] = presenter.sendFileMessage(request,
@@ -1511,15 +1515,13 @@ public class ChatActivity extends AppCompatActivity
                     public void onFinishImage(String json, ChatResponse<ResultImageFile> chatResponse) {
 
                         Log.e("SFM", "onFinishImage");
-                        url= chatResponse.getResult().getUrl();
-                        imageHashCode= chatResponse.getResult().getHashCode();
+
                     }
 
                     @Override
                     public void onFinishFile(String json, ChatResponse<ResultFile> chatResponse) {
 
                         Log.e("SFM", "onFinishFile");
-                        url= chatResponse.getResult().getUrl();
 
                     }
 
@@ -1988,18 +1990,25 @@ public class ChatActivity extends AppCompatActivity
 
             case 31: {
 
-                changePublicThreadToPrivate();
+                getAssistantHistory();
 
                 break;
             }
 
             case 32: {
 
+                changePublicThreadToPrivate();
+
+                break;
+            }
+
+            case 33: {
+
                 changePrivateThreadToPublic();
 
                 break;
             }
-            case 33: {
+            case 34: {
 
                 getHashTagHistory();
 
@@ -2065,20 +2074,29 @@ public class ChatActivity extends AppCompatActivity
     }
 
     private void registerAssistant() {
+        //63263 kheirkhah
+        ////63264 anvari
+
+        //63254 sajadi
+        //63256 amjadi
+        //63257 zhiani
         //invite list
-        Invitee invite = new Invitee("5739", InviteType.Constants.TO_BE_USER_CONTACT_ID);
+
+        //103187 nemati sandbox
+//        52987 sajadi 9063 sandbox
+       Invitee invite = new Invitee("63265", InviteType.Constants.TO_BE_USER_CONTACT_ID);
 
         //roles
         ArrayList<String> typeRoles = new ArrayList<>();
-        typeRoles.add(RoleType.Constants.READ_THREAD);
-        typeRoles.add(RoleType.Constants.EDIT_THREAD);
+        typeRoles.add(RoleType.Constants.REMOVE_ROLE_FROM_USER);
+
 
 
 
         List<AssistantVo> assistantVos = new ArrayList<>();
         AssistantVo assistantVo = new AssistantVo();
         assistantVo.setInvitees(invite);
-        assistantVo.setContactType("0");
+        assistantVo.setContactType("default");
         assistantVo.setRoles(typeRoles);
 
         assistantVos.add(assistantVo);
@@ -2090,14 +2108,32 @@ public class ChatActivity extends AppCompatActivity
 
     private void deactiveAssistant() {
 
+        //invite
+//        Invitee invite = new Invitee("52987", InviteType.Constants.TO_BE_USER_CONTACT_ID);
+        Invitee invite = new Invitee("63265", InviteType.Constants.TO_BE_USER_CONTACT_ID);
 
+
+        List<AssistantVo> assistantVos = new ArrayList<>();
+        AssistantVo assistantVo = new AssistantVo();
+        assistantVo.setInvitees(invite);
+        assistantVos.add(assistantVo);
+
+
+        DeActiveAssistantRequest request = new DeActiveAssistantRequest.Builder(assistantVos).build();
+        presenter.deActiveAssistant(request);
     }
 
     private void getAssistants() {
+        GetAssistantRequest request = new GetAssistantRequest.Builder().typeCode("default").setOffset(0).setCount(2).build();
 
+        presenter.getAssistants(request);
 
     }
 
+    private void getAssistantHistory(){
+        GetAssistantHistoryRequest request = new GetAssistantHistoryRequest.Builder().build();
+        presenter.getAssistantHistory(request);
+    }
 
 
     private void changePublicThreadToPrivate() {
@@ -2291,8 +2327,8 @@ public class ChatActivity extends AppCompatActivity
 
 
         Invitee[] invite = new Invitee[]{
-                new Invitee("103187", InviteType.Constants.TO_BE_USER_CONTACT_ID),
-                new Invitee("52987", InviteType.Constants.TO_BE_USER_CONTACT_ID),
+                new Invitee("5739", InviteType.Constants.TO_BE_USER_CONTACT_ID),
+                new Invitee("5740", InviteType.Constants.TO_BE_USER_CONTACT_ID),
         };
 
         String metac = getMetaData();
@@ -2388,7 +2424,6 @@ public class ChatActivity extends AppCompatActivity
 //        };
 
 
-
         /**
          *
          *
@@ -2425,9 +2460,9 @@ public class ChatActivity extends AppCompatActivity
         // add by user SSO_ID
 //                                invite.add(new Invitee(122, 1));  //user jiji
 //        invite.add(new Invitee("121", 1)); // user zizi
-        invite.add(new Invitee("52979", InviteType.Constants.TO_BE_USER_CONTACT_ID)); //anvari integ
-//        invite.add(new Invitee("120856", InviteType.Constants.TO_BE_USER_CONTACT_ID)); //nemati integ
-//        invite.add(new Invitee("63269", InviteType.Constants.TO_BE_USER_CONTACT_ID)); //pahlevani integ
+        invite.add(new Invitee("63272", InviteType.Constants.TO_BE_USER_CONTACT_ID)); //anvari integ
+        invite.add(new Invitee("120856", InviteType.Constants.TO_BE_USER_CONTACT_ID)); //nemati integ
+        invite.add(new Invitee("63269", InviteType.Constants.TO_BE_USER_CONTACT_ID)); //pahlevani integ
 //        invite.add(new Invitee("63268", InviteType.Constants.TO_BE_USER_CONTACT_ID));
 //        invite.add(new Invitee("80617", InviteType.Constants.TO_BE_USER_CONTACT_ID));
 //        invite.add(new Invitee("1", InviteType.Constants.TO_BE_USER_ID)); //amjadi
@@ -2447,7 +2482,8 @@ public class ChatActivity extends AppCompatActivity
 
 
         RequestCreateThread requestCreateThread = new RequestCreateThread
-                .Builder(ThreadType.Constants.NORMAL, invite)
+                .Builder(ThreadType.Constants.OWNER_GROUP, invite)
+                .title("Test Typing thread")
 //                .withDescription("Description created at "
 //                        + new Date().getTime())
 //                .withImage("URL")
@@ -2478,17 +2514,24 @@ public class ChatActivity extends AppCompatActivity
                 .count(25)
                 .order("desc") //.order("asc")
 //                .fromTime(new Date().getTime())
-             //   .toTime(new Date().getTime())
+                //   .toTime(new Date().getTime())
 //                .setMessageType(TextMessageType.Constants.POD_SPACE_PICTURE)
 //                .withNoCache()
                 .build();
 
         presenter.getHistory(request, null);
-
-
-
-
     }
+
+//    public void getHashTagList() {
+//        RequestGetHashTagList request = new RequestGetHashTagList
+//                .Builder(TEST_THREAD_ID)
+//                .offset(0)
+//                .count(25)
+//                .setHashtag("ahmad")
+//                .build();
+//
+//        presenter.getHashTagLIst(request, null);
+//    }
 
     public void getHashTagHistory() {
         RequestGetHashTagList request = new RequestGetHashTagList
@@ -2734,8 +2777,6 @@ public class ChatActivity extends AppCompatActivity
                     Toast.makeText(getApplicationContext(), "Finish Upload", Toast.LENGTH_SHORT).show();
                     percentage.setTextColor(getResources().getColor(R.color.colorAccent));
                     percentage.setText("100");
-                    url=chatResponse.getResult().getUrl();
-                    imageHashCode= chatResponse.getResult().getHashCode();
                 });
 
             }