+++ app/src/main/java/com/example/chat/application/chatexample/ConstantSample.java
@@ -36,6 +36,8 @@ public class ConstantSample {
             "Register Assistant",
             "deactive Assistant",
             "get Assistants",
+            "get Assistants History",
+            "get Assistants",
             "Convert Public Thread To Private",
             "Convert Private Thread To Public",
             "Get Hash Tag List",